feat: improve accuracy of gpt.sh script output
The gpt.sh script was significantly improved in this commit. Firstly, the command in line 2 was modified using "git diff". Secondly, a while loop was added to read each line of the output and use a case statement to determine whether a line was added or deleted using "+" and "-" respectively. Finally, the command in line 2 was refactored to use the while loop with a grandparent revision specified for greater accuracy.

---

Changes to file scripts/gpt.sh:

- Modified the command in line 2 using "git diff".
- Added a while loop to read each line of the output and use a case statement to determine whether a line was added or deleted using "+" and "-" respectively.
- Refactored the command in line 2 to use the while loop with a grandparent revision specified for greater accuracy.